Split mana.py into modules
