--- usr/local/sbin/mana
@@ -1,134 +1,0 @@
-#!/bin/sh
-
-# mana: node management tool
-
-wait_lnd_stopped() {
-    WAITLNDSTOPPED=1
-    while [ "$(docker ps | grep -c compose_lnd_1)" = 1 ] && [ $WAITLNDSTOPPED -lt 10 ];
-    do
-        echo "Waiting for LND to terminate..."
-        WAITLNDSTOPPED=$((WAITLNDSTOPPED+1))
-        sleep 2
-    done
-}
-
-wait_bitcoind_stopped() {
-    WAITBTCSTOPPED=1
-    while [ "$(docker ps | grep -c compose_bitcoind_1)" = 1 ] && [ $WAITBTCSTOPPED -lt 10 ];
-    do
-        echo "Waiting for bitcoind to terminate..."
-        WAITBTCSTOPPED=$((WAITBTCSTOPPED+1))
-        sleep 2
-    done
-}
-
-cmd_exists() {
-  $(command -v ${1} 2>&1 1>/dev/null;)
-  ret=$?
-  echo $ret
-  return $ret
-}
-
-install_git() {
-  if [ "$(cmd_exists git)" -ne "0" ]; then
-    apk update && \
-    apk add git
-  fi
-}
-
-if [ "$1" = "info" ]; then
-  # bitcoind and lnd information
-  echo "bitcoind info:"
-  docker exec -it compose_bitcoind_1 bitcoin-cli -getinfo
-  echo "lnd info:"
-  docker exec -it compose_lnd_1 lncli getinfo
-elif [ "$1" = "devtools" ]; then
-  apk update && \
-  apk add tmux sudo git rsync htop iotop nmap nano glances
-elif [ "$1" = "source" ]; then
-  install_git
-  if [ ! -d /home/lncm/pi-factory ]; then
-    cd /home/lncm || exit
-    git clone https://github.com/lncm/pi-factory.git
-  fi
-  if [ -d /home/lncm/pi-factory ]; then
-    cd /home/lncm/pi-factory || exit
-    git pull
-  fi
-elif [ "$1" = "full-upgrade" ]; then
-  # Replaces entire FAT contents, 
-  # as if we installed from freshly burned SD card
-  mana source
-  cd /home/lncm/pi-factory
-  git pull
-  sh make_upgrade.sh
-elif [ "$1" = "upgrade" ]; then
-  # Regenerate box.apkovl.tar.gz and mark SD as uninstalled
-  mana source
-  cd /home/lncm/pi-factory
-  git pull
-  cp -v /etc/wpa_supplicant/wpa_supplicant.conf etc/wpa_supplicant/wpa_supplicant.conf
-  ./make_apkovl.sh
-  mount -o remount,rw /dev/mmcblk0p1 /media/mmcblk0p1
-  cp -v box.apkovl.tar.gz /media/mmcblk0p1/
-  rm -v /media/mmcblk0p1/installed
-  mount -o remount,ro /dev/mmcblk0p1 /media/mmcblk0p1
-  echo "Done"
-  echo "Please reboot to upgrade your box"
-elif [ "$1" = "diff" ]; then
-  install_git
-  if [ ! -d /home/lncm/pi-factory ]; then
-    cd /home/lncm || exit
-    git clone https://github.com/lncm/pi-factory.git
-  fi
-  if [ -d /home/lncm/pi-factory ]; then
-    cd /home/lncm/pi-factory || exit
-    git pull || exit
-    echo "Writing .diff logs to /home/lncm"
-    diff -r /etc /home/lncm/pi-factory/etc | tee /home/lncm/etc.diff
-    diff -r /usr /home/lncm/pi-factory/usr | tee /home/lncm/usr.diff
-    diff -r /home /home/lncm/pi-factory/home | tee /home/lncm/home.diff
-  fi
-elif [ "$1" = "unlock" ]; then
-  # unlock lnd wallet
-  docker exec -it compose_lnd_1 lncli unlock
-elif [ "$1" = "create" ]; then
-  # Use create wallet script. This will either import an existing seed (or our own generated one), or use LND to create one. It will also create a password either randomly or use an existing password provided)
-  /usr/local/sbin/lncm-createwallet.py
-elif [ "$1" = "start" ]; then
-  service docker-compose start
-elif [ "$1" = "stop" ]; then
-  echo "Attempting clean shutdown of bitcoind and lnd nodes"
-  docker exec -it compose_lnd_1 lncli stop
-  docker exec -it compose_bitcoind_1 bitcoin-cli stop
-  wait_lnd_stopped
-  wait_bitcoind_stopped
-  # Stop docker compose
-  service docker-compose stop
-elif [ "$1" = "restart" ]; then
-  echo "Attempting clean shutdown of bitcoind and lnd nodes"
-  docker exec -it compose_lnd_1 lncli stop
-  docker exec -it compose_bitcoind_1 bitcoin-cli stop
-  wait_lnd_stopped
-  wait_bitcoind_stopped
-  service docker-compose restart
-elif [ "$1" = "btclogs" ]; then
-  # slow, robust method
-  docker logs -f compose_bitcoind_1
-elif [ "$1" = "btclog" ]; then
-  tail -f /media/volatile/volatile/bitcoin/debug.log
-elif [ "$1" = "lndlogs" ]; then
-  # slow, robust method
-  docker logs -f compose_lnd_1
-elif [ "$1" = "lndlog" ]; then
-  tail -f /media/volatile/volatile/lnd/logs/bitcoin/mainnet/lnd.log
-elif [ "$1" = "temp" ]; then
-  # CPU temperature
-  cpu=$(cat /sys/class/thermal/thermal_zone0/temp)
-  echo "CPU: $((cpu/1000))C"
-else
-  echo "Mana: node management"
-  echo
-  echo "Must pass one of:"
-  echo "info, unlock, create, stop, start, restart, btclog, lndlog, temp"
-fi
